Introduce Libvirt::Hypervisor.
The goal of this commit is to introduce common api
to manipulate libvirt objects. In longer run it will
allow to introduce a lxc domain and other lxc types
without duplicating large chunks of code.

Moreover encapsulate libvirt-ruby so it'll
be possible to replace it with better library without rewritting
all providers. Other aren't rewritten to use the hypervisor object though.