Add support for LXC domains.
Now it's possible to create and run LXC containers.

Due to restrictions in the libvirt-lxc driver the
domain must be defined along with filesystems and network.
Neither a filesystem nor a network can be attached runtime.
It makes the LXC provider less elastic than KVM but it covers
the most frequent usecase.

While this provider requires a filesystem and a network, the
appropriate attributes has been added to the resource definition.
The are ignored by KVM though. On the other hand the lxc provider
ignores the arch attribute required by KVM.
Maybe the domain resource should be split into lxc_domain and kvm_domain.